requirements.txt
aiofiles==0.6.0
aiogram==2.12.1
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==20.3.0
auto-py-to-exe==2.7.8
Babel==2.9.0
beautifulsoup4==4.9.3
bottle==0.12.18
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
ci-info==0.2.0
click==7.1.2
comtypes==1.1.7
configobj==5.0.6
configparser==5.0.2
decorator==4.4.2
Eel==0.12.4
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==1.1.2
Flask-WTF==0.14.3
frontend==0.0.3
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
h11==0.12.0
httplib2==0.19.0
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
multidict==5.1.0
nest-asyncio==1.4.3
networkx==2.5
nibabel==3.2.1
nipype==1.6.0
numpy==1.19.4
packaging==20.9
pandas==1.2.3
pathlib==1.0.1
pdfminer==20191125
pefile==2019.4.18
Pillow==7.2.0
prov==2.0.0
pycparser==2.20
pycryptodome==3.10.1
pydot==1.4.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyMuPDF==1.18.10
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
PyQt5==5.13.2
PyQt5-sip==12.8.1
pyqt5-tools==5.13.2.1.6rc1
pyqtgraph==0.11.0
pyTelegramBotAPI==3.7.7
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pyttsx3==2.90
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
pyxnat==1.4
rdflib==5.0.0
requests==2.25.1
schedule==0.6.0
scipy==1.6.2
simplejson==3.17.2
six==1.15.0
snake-on-pygame==1.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
starlette==0.14.2
tika==1.24
traits==6.2.0
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.4
vk-api==11.9.1
Werkzeug==1.0.1
whichcraft==0.6.1
WTForms==2.3.3
xmltodict==0.12.0
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.1.2
